feat(shortcuts): Add A2A and MCP shortcuts for listing connected servers and their statuses
